Arregle unos warnings con respecto al tipo de dato (antes era int ahora es long para poder imprimirlo como otro valor hex) que se utiliza para guardar los valores de memoria leidos

+++ Userland/SampleCodeModule/api.c
@@ -45,7 +45,7 @@ int getDay(){
 }
 
 
-void readMem(int * mem , long * dir){
+void readMem(long * mem , long * dir){
     readMemAsm(mem,dir);
 }
